Refactor all short url request operations to a service
The intention of this commit is to move all the business logic from
hooks to a service.

This allows a developer to see, in a single place, what happens when
each operation is performed, rather than having the behaviour of the
application implemented as 'side effects' that have to be traced in
multiple places - the controller, the model, and its associated
redirect.